fix: 토너먼트 로직 롤백 (#456)
